+++ src/lib.rs
@@ -1,5 +1,5 @@
 pub(crate) mod classify;
-pub(crate) mod config;
+pub mod config;
 pub mod containers;
 pub(crate) mod cpu;
 pub(crate) mod gpu;